make Lockfile an interface instead of abstract class
