+++ packages/cli-node/src/pacman/yarn/YarnLockfile.ts
@@ -16,7 +16,8 @@
 
 import { parseSyml, stringifySyml } from '@yarnpkg/parsers';
 import { stringify as legacyStringifyLockfile } from '@yarnpkg/lockfile';
-import { Lockfile, LockfileData, LockfileQueryEntry } from '../Lockfile';
+import { Lockfile, LockfileDiff } from '../Lockfile';
+import crypto from 'node:crypto';
 
 const ENTRY_PATTERN = /^((?:@[^/]+\/)?[^@/]+)@(.+)$/;
 
@@ -44,7 +45,30 @@ const SPECIAL_OBJECT_KEYS = [
   `binaries`,
 ];
 
-export class YarnLockfile extends Lockfile {
+type LockfileQueryEntry = {
+  range: string;
+  version: string;
+  dataKey: string;
+};
+
+type LockfileData = {
+  [entry: string]: {
+    version: string;
+    resolved?: string;
+    integrity?: string /* old */;
+    checksum?: string /* new */;
+    dependencies?: { [name: string]: string };
+    peerDependencies?: { [name: string]: string };
+  };
+};
+
+export class YarnLockfile implements Lockfile {
+  private constructor(
+    private packages: Map<string, LockfileQueryEntry[]>,
+    private data: LockfileData,
+    private readonly legacy: boolean,
+  ) {}
+
   static parse(content: string) {
     const legacy = LEGACY_REGEX.test(content);
 
@@ -84,12 +108,139 @@ export class YarnLockfile extends Lockfile {
     return new YarnLockfile(packages, data, legacy);
   }
 
-  private constructor(
-    packages: Map<string, LockfileQueryEntry[]>,
-    data: LockfileData,
-    private readonly legacy: boolean,
-  ) {
-    super(packages, data);
+  get(name: string): LockfileQueryEntry[] | undefined {
+    return this.packages.get(name);
+  }
+
+  keys(): IterableIterator<string> {
+    return this.packages.keys();
+  }
+  createSimplifiedDependencyGraph(): Map<string, Set<string>> {
+    const graph = new Map<string, Set<string>>();
+
+    for (const [name, entries] of this.packages) {
+      const dependencies = new Set(
+        entries.flatMap(e => {
+          const data = this.data[e.dataKey];
+          return [
+            ...Object.keys(data?.dependencies ?? {}),
+            ...Object.keys(data?.peerDependencies ?? {}),
+          ];
+        }),
+      );
+      graph.set(name, dependencies);
+    }
+
+    return graph;
+  }
+  diff(otherLockfile: YarnLockfile): LockfileDiff {
+    const diff = {
+      added: new Array<{ name: string; range: string }>(),
+      changed: new Array<{ name: string; range: string }>(),
+      removed: new Array<{ name: string; range: string }>(),
+    };
+
+    // Keeps track of packages that only exist in this lockfile
+    const remainingOldNames = new Set(this.packages.keys());
+
+    for (const [name, otherQueries] of otherLockfile.packages) {
+      remainingOldNames.delete(name);
+
+      const thisQueries = this.packages.get(name);
+      // If the packages don't exist in this lockfile, add all entries
+      if (!thisQueries) {
+        diff.removed.push(...otherQueries.map(q => ({ name, range: q.range })));
+        continue;
+      }
+
+      const remainingOldRanges = new Set(thisQueries.map(q => q.range));
+
+      for (const otherQuery of otherQueries) {
+        remainingOldRanges.delete(otherQuery.range);
+
+        const thisQuery = thisQueries.find(q => q.range === otherQuery.range);
+        if (!thisQuery) {
+          diff.removed.push({ name, range: otherQuery.range });
+          continue;
+        }
+
+        const otherPkg = otherLockfile.data[otherQuery.dataKey];
+        const thisPkg = this.data[thisQuery.dataKey];
+        if (otherPkg && thisPkg) {
+          const thisCheck = thisPkg.integrity || thisPkg.checksum;
+          const otherCheck = otherPkg.integrity || otherPkg.checksum;
+          if (thisCheck !== otherCheck) {
+            diff.changed.push({ name, range: otherQuery.range });
+          }
+        }
+      }
+
+      for (const thisRange of remainingOldRanges) {
+        diff.added.push({ name, range: thisRange });
+      }
+    }
+
+    for (const name of remainingOldNames) {
+      const queries = this.packages.get(name) ?? [];
+      diff.added.push(...queries.map(q => ({ name, range: q.range })));
+    }
+
+    return diff;
+  }
+  getDependencyTreeHash(startName: string): string {
+    if (!this.packages.has(startName)) {
+      throw new Error(`Package '${startName}' not found in lockfile`);
+    }
+
+    const hash = crypto.createHash('sha1');
+
+    const queue = [startName];
+    const seen = new Set<string>();
+
+    while (queue.length > 0) {
+      const name = queue.pop()!;
+
+      if (seen.has(name)) {
+        continue;
+      }
+      seen.add(name);
+
+      const entries = this.packages.get(name);
+      if (!entries) {
+        continue; // In case of missing optional peer dependencies
+      }
+
+      hash.update(`pkg:${name}`);
+      hash.update('\0');
+
+      // TODO(Rugvip): This uses the same simplified lookup as createSimplifiedDependencyGraph()
+      //               we could match version queries to make the resulting tree a bit smaller.
+      const deps = new Array<string>();
+      for (const entry of entries) {
+        // We're not being particular about stable ordering here. If the lockfile ordering changes, so will likely hash.
+        hash.update(entry.version);
+
+        const data = this.data[entry.dataKey];
+        if (!data) {
+          continue;
+        }
+
+        const checksum = data.checksum || data.integrity;
+        if (checksum) {
+          hash.update('#');
+          hash.update(checksum);
+        }
+
+        hash.update(' ');
+
+        deps.push(...Object.keys(data.dependencies ?? {}));
+        deps.push(...Object.keys(data.peerDependencies ?? {}));
+      }
+
+      queue.push(...new Set(deps));
+    }
+
+    return hash.digest('hex');
   }
 
   toString() {
